// sistema/src/components/shared/Interfaces.ts
export interface imagenUpladvalues {
  visualizacion: string
  file: File
  filename: string
}

export interface contenidoSeccionValues {
  id: string
  titulo: string
  tipo: string
  contenido: string | imagenUpladvalues[]
}

export interface temarioValues {
  id: string
  titulo: string
  clase: contenidoSeccionValues[]
}